+++ client/src/components/layout/Sidebar.jsx
@@ -1,7 +1,7 @@
 import { useContext } from 'react';
 import { NavLink, Link } from 'react-router-dom';
-import { Context } from '../context/Context';
-import ToggleThemeButton from './ToggleThemeButton';
+import { Context } from '../../context/Context';
+import ToggleThemeButton from '../ToggleThemeButton';
 
 import logo from '../../public/assets/logo.svg';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
